Refactor: Move classes to new package
